+++ googleapis_auth/test/iam_signer_test.dart
@@ -9,6 +9,7 @@ library;
 
 import 'dart:convert';
 import 'package:googleapis_auth/src/iam_signer.dart';
+import 'package:googleapis_auth/src/version.dart';
 import 'package:http/http.dart';
 import 'package:http/testing.dart';
 import 'package:test/test.dart';
@@ -18,7 +19,10 @@ void main() {
     final client = MockClient((request) async {
       if (request.url.path.contains('signBlob')) {
         expect(request.url.toString(), contains('test-email%40example.com'));
-        expect(request.headers['x-goog-api-client'], isNotNull);
+        expect(
+          request.headers[xGoogApiClientHeader],
+          xGoogApiClientHeaderValue,
+        );
         final body = jsonDecode(request.body) as Map<String, dynamic>;
         expect(body['payload'], isNotNull);
         return Response(